[FIX] 오리지널이름 버그 해결2

+++ src/main/java/MeshX/HypeLink/head_office/item/model/dto/request/CreateItemImageReq.java
@@ -11,7 +11,7 @@ import lombok.ToString;
 @ToString
 @EqualsAndHashCode
 public class CreateItemImageReq {
-    private Long id;  // ← 이미지 ID 추가
+    private Integer id;  // ← 이미지 ID 추가
     private String originalFilename;
     private Long fileSize;
     private String contentType;
